Add the ability to do conversion in case images and annotations lies at the same folder
